+++ src/components/dialog/PostDetailDialog.tsx
